Arranging the entire router of puzzles to be a clean file, and work with a controller
