+++ Server/DAL/DB/riddelsDB.txt
@@ -21,9 +21,9 @@
     "id": 102
   },
   {
-    "name": "g",
-    "taskDescription": "h",
-    "correctAnswer": "h",
+    "name": "f",
+    "taskDescription": "f",
+    "correctAnswer": "f",
     "difficulty": "hard",
     "id": 103
   }
